+++ src/main/java/refinedstorage/tile/TileMachine.java
@@ -1,7 +1,6 @@
 package refinedstorage.tile;
 
 import io.netty.buffer.ByteBuf;
-import net.minecraft.block.Block;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.util.math.BlockPos;
 import refinedstorage.block.BlockMachine;
@@ -13,10 +12,8 @@ public abstract class TileMachine extends TileBase implements INetworkTile, IRed
     protected RedstoneMode redstoneMode = RedstoneMode.IGNORE;
     protected TileController controller;
 
-    private Block machineBlock;
-
     public void onConnected(TileController controller) {
-        if (worldObj != null && worldObj.getBlockState(pos).getBlock() == machineBlock) {
+        if (worldObj != null && worldObj.getBlockState(pos).getBlock() == getBlockType()) {
             markDirty();
 
             this.connected = true;
@@ -27,7 +24,7 @@ public abstract class TileMachine extends TileBase implements INetworkTile, IRed
     }
 
     public void onDisconnected() {
-        if (worldObj != null && worldObj.getBlockState(pos).getBlock() == machineBlock) {
+        if (worldObj != null && worldObj.getBlockState(pos).getBlock() == getBlockType()) {
             markDirty();
 
             this.connected = false;
@@ -43,18 +40,9 @@ public abstract class TileMachine extends TileBase implements INetworkTile, IRed
 
     @Override
     public void update() {
-        if (worldObj == null) {
-            super.update();
-            return;
-        }
-
-        if (ticks == 0) {
-            machineBlock = worldObj.getBlockState(pos).getBlock();
-        }
-
         super.update();
 
-        if (!worldObj.isRemote && isConnected()) {
+        if (worldObj != null && !worldObj.isRemote && isConnected()) {
             updateMachine();
         }
     }